03_ewkt_i_here/requirements.txt
geopandas
sqlalchemy
psycopg2-binary
python-dotenv
geoalchemy2
